Properly handle scalar repeated fields (fixes #13)
Note that for repeated fields is_set has to always be true, as otherwise
the rules in current OSI are non-sensical. We should introduce new
primitives to check for at least one entry in repeated fields for cases
where we do want to have that. Similarly this enhances robustness
against illegal rules, like the refers_to rule on non-identifier type
messages.

Signed-off-by: Pierre R. Mai <pmai@pmsf.de>